m219 血量突破int上限+全模组K/M/B/T紧凑数字:属性上限1e9→1e15;8处BOSS血条‖(int)强转(卡死2147483647的真凶)全改(long);新建NumFmt统一四处显示(血条/HUD/成长面板/装备介绍)

+++ src/main/java/com/yongye/entity/RedSpiderEntity.java
@@ -59,7 +59,7 @@ public class RedSpiderEntity extends SpiderEntity implements GeoEntity {
             float max = this.getMaxHealth();
             // m187:血量数字嵌入血条名(‖当前/最大)→ 客户端解析显示
             this.bossBar.setName(this.getType().getName().copy().formatted(Formatting.RED)
-                    .append(Text.literal("\u2016" + (int)this.getHealth() + "/" + (int)max)));
+                    .append(Text.literal("\u2016" + (long)this.getHealth() + "/" + (long)max)));
             this.bossBar.setPercent(max > 0 ? Math.max(0f, Math.min(1f, this.getHealth() / max)) : 0f);
         }
     }
